requirements.txt
Django==4.2
mysqlclient==2.2.5
drf-spectacular==0.27.2
djangorestframework==3.14.0